requirements.txt
numpy==1.22.3
Pillow==9.4.0
streamlit==1.18.1
tensorflow==2.11.0
tensorflow_macos==2.9.2
